User/spencer aherne/goblin enemy (#5)
* Goblin setup

Basic start for the goblin enemy, including sprite and some scripts. Goblin is currently able to detect if something is between it and the player. Need to make it move to be able to see player though.

* Goblin movement and basic attack

Goblin now moves until direct line of sight player, and the script has a very basic attack mechanic, though not working the way I plan. Also consolidated goblin and goblinMovement scripts into one.

* Goblin comments

Mostly just comments on the goblin script to try and figure out how to attack, so they can be looked over by others.

* Removed .csproj and .sln from gitignore

Files were needed for others to open the projects.

* restored gitignore

* Goblin and gitIgnore

Started getting the random attack arc working on the Goblin enemy.

Not sure what happened with gitignore. I'm guessing that when I opened it with notepad and renamed it a bunch of stuff moved around. Hopefully this won't be a problem.

* Hopefully fixed my gitignore file

* Goblin attack and GitIgnore

Goblin attack is almsot complete, just need to figure out how to properly rotate the arrow towards the attack point. Also tried to do more to fix gitignore, will see if this sticks or not.

* trying to fix arrow sprite

Arrow sprite spawns at odd rotations based on where the goblin is in the scene. Still not working as intended.

* Fixing GitIgnore Maybe

Deleteing things that gitignore should ignore, hoping this will solve the problem and not break my project.

* Update .gitignore

* Change projectile

Changed the goblins projectile from an arrow to a stone to avoid rotation issues I couldn't solve. Also cleaned up goblin script a bit following the projectile changes.

* Do and Take Damage

Added and adjusted scripts so the player could damage the goblin as well as the skeleton, and so the goblin could damage the player properly.

* Update Goblin.cs

Just a quick clean up of Goblin.cs and an added note of potential mechanic changes, depending on how I want it to work in the future.