Walls, doors, and events
Added proper walls to the game, as well as adjusting to doors code to behave solid while room contains enemies. Adjusted several prefabs to include proper walls, colliders, and triggers as required, as well as positions for easier room making in the future. Also added an event for when a room is cleared of enemies, but hasn't been tested.