Fixed Chest Loot/Spawn, Boss Door, and Adjusted Player Aiming
Made chest spawn loot in an area where it can be picked up, as well as made chests properly spawn properly if they are rolled in the room clear loot table.

Boss door now opens when the boss is killed. Was forgetting to remove boss from list of enemies when defeated.

Did some work on the player aiming system, I think it's not too different from when it started, but I feel like it's a bit better than the previous system.